% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geocorr_crosswalk.R
\name{get_geocorr_crosswalk}
\alias{get_geocorr_crosswalk}
\title{Obtain a Geocorr22 Crosswalk}
\usage{
get_geocorr_crosswalk(
  source_geography,
  target_geography,
  weight = c("population", "housing", "land"),
  cache = NULL,
  overwrite_cache = FALSE
)
}
\arguments{
\item{source_geography}{Source geography type.}

\item{target_geography}{Target geography type}

\item{weight}{Weighting variable. One of c("population", "housing", "land").}

\item{cache}{NULL by default. If not NULL, the path at which to cache the results.}

\item{overwrite_cache}{FALSE by default. If TRUE, the existing cached crosswalk will be replaced.}
}
\value{
A dataframe representing the requested Geocorr22 crosswalk for all 51 states and Puerto Rico. Depending on the desired geographies, some fields may not be included.
\describe{
\item{state_fips}{A two-digit identified for the state (or DC/PR), if applicable}
\item{state_abbreviation}{A two-character abbreviation for the state (or DC/PR), if applicable}
\item{source_geoid}{A unique identifier for the source geography}
\item{target_geoid}{A unique identifier for the target geography}
\item{source_geography_name}{The name of the source geography}
\item{target_geography_name}{The name of the target geography}
\item{allocation_factor_source_to_target}{The weight to interpolate values from the source geography to the target geography}
\item{allocation_factor_target_to_source}{The weight to interpolate values from the source geography to the target geography}
\item{population_2020}{The estimated overlap in population, if applicable}
\item{housing_2020}{The estimated overlap in housing units, if applicable}
\item{land_area_sqmi}{The overlap in land area, if applicable}
\item{weighting_factor}{The attribute used to calculate allocation factors (one of population, housing, land)}
}
}
\description{
Query Geocorr22 (https://mcdc.missouri.edu/applications/geocorr2022.html) for
a crosswalk between two geographies for all 51 states and Puerto Rico.
}
\details{
Note: this function is under development but does not yet support all of the geographies
supported by Geocorr. Currently this includes:
c("place", "county", "tract", "blockgroup", "zcta", "puma22", "cd119", "cd118").
}
