% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhgis_crosswalk.R
\name{get_nhgis_crosswalk}
\alias{get_nhgis_crosswalk}
\title{Get NHGIS Geographic Crosswalk}
\usage{
get_nhgis_crosswalk(
  source_year,
  source_geography,
  target_year,
  target_geography,
  api_key = NULL
)
}
\arguments{
\item{source_year}{Character or numeric. Year of the source geography one of c(1990, 2000, 2010, 2020).}

\item{source_geography}{Character. Source geography name. One of c("block", "block group part", "tract").}

\item{target_year}{Character or numeric. Year of the target geography, one of c(1990, 2000, 2010, 2020).}

\item{target_geography}{Character. Target geography name. One of c("block", "block group part", "tract", "county").}

\item{api_key}{Character. NULL by default, in which case the function looks for an \code{IPUMS_API_KEY} environment variable.}
}
\value{
A data frame containing the crosswalk between the specified geographies.
}
\description{
Retrieves a geographic crosswalk from the IPUMS NHGIS API based on user-specified
source and target geographies and years. Accepts various spellings and formats
for geography names.
}
\keyword{internal}
