% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhgis_crosswalk.R
\name{get_nhgis_crosswalk}
\alias{get_nhgis_crosswalk}
\title{Get NHGIS Geographic Crosswalk}
\usage{
get_nhgis_crosswalk(
  source_year,
  source_geography,
  target_year,
  target_geography,
  download_directory,
  use_cache = FALSE,
  api_key = NULL
)
}
\arguments{
\item{source_year}{Character or numeric. Year of the source geography one of c(1990, 2000, 2010, 2020).}

\item{source_geography}{Character. Source geography name. One of c("block", "block group", "tract").}

\item{target_year}{Character or numeric. Year of the target geography, one of c(1990, 2000, 2010, 2020).}

\item{target_geography}{Character. Target geography name. One of c("block", "block group", "tract", "place", county", "urban_area", "zcta", "puma", "core_based_statistical_area").}

\item{download_directory}{File path. Where to download the crosswalk to.}

\item{use_cache}{FALSE by default. If TRUE, read in an already-downloaded crosswalk stored in the \code{download_directory}, if such a file exists.}

\item{api_key}{Character. NULL by default, in which case the function looks for an \code{IPUMS_API_KEY} environment variable.}
}
\value{
A data frame containing the crosswalk between the specified geographies. Data are tidy-formatted, with each observation reflecting a unique source-target-weighting factor combination. Note that all (typically two or three) available weighting factors are returned.

#' @return A dataframe representing the requested Geocorr22 crosswalk for all 51 states and Puerto Rico. Depending on the desired geographies, some fields may not be included.
\describe{
\item{source_geoid}{A unique identifier for the source geography}
\item{target_geoid}{A unique identifier for the target geography}
\item{source_geography_name}{The name of the source geography}
\item{target_geography_name}{The name of the target geography}
\item{source_year}{The year of the source geography}
\item{target_year}{The year of the target geography}
\item{allocation_factor_source_to_target}{The weight to interpolate values from the source geography to the target geography}
\item{weighting_factor}{The attribute used to calculate allocation factors}
}
}
\description{
Retrieves a geographic crosswalk from the IPUMS NHGIS API based on user-specified
source and target geographies and years. Use \code{list_nhgis_crosswalks()} to view valid
parameter combinations.
}
\details{
Note: for the moment, this function does not support block group part crosswalks.
}
