% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhgis_crosswalk.R
\name{list_nhgis_crosswalks}
\alias{list_nhgis_crosswalks}
\title{List supported NHGIS crosswalks}
\usage{
list_nhgis_crosswalks()
}
\value{
A tibble with columns:
\itemize{
\item source_year: Year of the source geography
\item source_geography: Source geography name
\item target_year: Year of the target geography
\item target_geography: Target geography name
}
}
\description{
Returns a tibble of all available NHGIS geographic crosswalks with their
corresponding parameters that can be used with get_nhgis_crosswalk().
}
